Add /history page with basic rendering for history entries, and refactor
