+++ src/main.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import { AsideNav } from "./components/AsideNav/index.tsx";
-import { AppRoutes } from "./AppRoutes.tsx";
 import { AppProvider } from "./context/index.tsx";
+import { AppRoutes } from "./pages/AppRoutes.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
